Moved sessionStorage utils to storage file. Simplified context provider to not need to redirect as managed by ProtectedRoutes. Added truthy constraint to username and password on login page to avoid needlessly requesting to API.
