+++ web/src/data/api.tsx
@@ -31,17 +31,3 @@ export const decodeUserAuth = (tokens: Tokens): AuthData => {
     lastName: idTokenPayload.lastname,
   };
 };
-
-export const setSessionUser = (user: AuthData): void => {
-  sessionStorage.setItem("trace_user", JSON.stringify(user));
-};
-
-export const getSessionUser = (): AuthData | null => {
-  const item = sessionStorage.getItem("trace_user");
-  if (item === null) return null;
-  return JSON.parse(item) as AuthData;
-};
-
-export const removeSessionUser = (): void => {
-  sessionStorage.removeItem("trace_user");
-};
