+++ web/tsconfig.json
@@ -15,22 +15,22 @@
     "noEmit": true,
     "jsx": "react-jsx",
 
-	"allowUnreachableCode": false,
-	"allowUnusedLabels": false,
-	"alwaysStrict": false,
-	//"exactOptionalPropertyTypes": true, // Incompatible with Drawer from @mui/material
-	"noImplicitAny": true,
-	"noImplicitOverride": true,
-	"noImplicitReturns": true,
-	"noImplicitThis": true,
-	"noPropertyAccessFromIndexSignature": true,
-	"noUncheckedIndexedAccess": true,
-	"strictBindCallApply": true,
-	"strictFunctionTypes": true,
-	"strictNullChecks": true,
-	"strictPropertyInitialization": true,
-	"useUnknownInCatchVariables": true,
-	"allowJs": false,
+    "allowUnreachableCode": false,
+    "allowUnusedLabels": false,
+    "alwaysStrict": false,
+    //"exactOptionalPropertyTypes": true, // Incompatible with Drawer from @mui/material
+    "noImplicitAny": true,
+    "noImplicitOverride": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "noPropertyAccessFromIndexSignature": true,
+    "noUncheckedIndexedAccess": true,
+    "strictBindCallApply": true,
+    "strictFunctionTypes": true,
+    "strictNullChecks": true,
+    "strictPropertyInitialization": true,
+    "useUnknownInCatchVariables": true,
+    "allowJs": false,
 
     /* Linting */
     "strict": true,
